requirements.txt
discord.py
PyNaCl
wavelink